+++ engine/engine_test.go
@@ -10,9 +10,9 @@ import (
 	"github.com/runabol/tork"
 	"github.com/runabol/tork/datastore"
 	"github.com/runabol/tork/input"
-	"github.com/runabol/tork/internal/runtime"
 	"github.com/runabol/tork/mount"
 	"github.com/runabol/tork/mq"
+	"github.com/runabol/tork/runtime"
 	"github.com/stretchr/testify/assert"
 )
