+++ internal/coordinator/coordinator_test.go
@@ -14,9 +14,9 @@ import (
 	"github.com/runabol/tork/middleware/task"
 	"github.com/runabol/tork/mq"
 
-	"github.com/runabol/tork/internal/runtime/docker"
 	"github.com/runabol/tork/internal/uuid"
 	"github.com/runabol/tork/internal/worker"
+	"github.com/runabol/tork/runtime/docker"
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/yaml.v3"
 )
