+++ internal/worker/api_test.go
@@ -6,8 +6,8 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/runabol/tork/internal/runtime/docker"
 	"github.com/runabol/tork/mq"
+	"github.com/runabol/tork/runtime/docker"
 	"github.com/stretchr/testify/assert"
 )
