+++ internal/worker/worker.go
@@ -14,8 +14,8 @@ import (
 	"github.com/runabol/tork/middleware/task"
 	"github.com/runabol/tork/mq"
 
-	"github.com/runabol/tork/internal/runtime"
 	"github.com/runabol/tork/internal/syncx"
+	"github.com/runabol/tork/runtime"
 
 	"github.com/runabol/tork/internal/uuid"
 )
